Move stratify-itunes into Rails app
