--- stratify-itunes/.gitignore
@@ -1,3 +1,0 @@
-*.gem
-.bundle
-pkg/*
