--- stratify-itunes/lib/stratify-itunes.rb
@@ -1,6 +1,0 @@
-require 'stratify-base'
-
-require 'stratify-itunes/activity'
-require 'stratify-itunes/collector'
-require 'stratify-itunes/engine'
-require 'stratify-itunes/version'
