--- stratify-itunes/lib/stratify-itunes/engine.rb
@@ -1,9 +1,0 @@
-require 'rails'
-
-module Stratify
-  module ITunes
-    class Engine < Rails::Engine
-      engine_name :stratify_itunes
-    end
-  end
-end
