--- stratify-itunes/lib/stratify-itunes/version.rb
@@ -1,5 +1,0 @@
-module Stratify
-  module ITunes
-    VERSION = "0.1.4.2"
-  end
-end
