+++ stratify-rails/lib/stratify/itunes.rb
@@ -1,0 +1,2 @@
+require 'stratify/itunes/activity'
+require 'stratify/itunes/collector'
