+++ stratify-rails/lib/stratify/itunes/assets/images/itunes-icon-16.png
