+++ stratify-rails/lib/stratify/itunes/assets/images/itunes-icon-24.png
