+++ stratify-rails/lib/stratify/itunes/presenter.rb
