+++ stratify-rails/spec/fixtures/iTunes Music Library.xml
