+++ stratify-rails/spec/lib/stratify/itunes/presenter_spec.rb
